Refactor State into a separate module
